[#9]refactor: 패키지 구조 병합(login, signup)
