+++ src/main/java/soma/haeya/edupi_user/client/MemberApiClient.java
@@ -1,11 +1,11 @@
-package soma.haeya.edupi_user.login.client;
+package soma.haeya.edupi_user.client;
 
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.service.annotation.HttpExchange;
 import org.springframework.web.service.annotation.PostExchange;
-import soma.haeya.edupi_user.login.domain.Member;
-import soma.haeya.edupi_user.login.dto.MemberLoginRequest;
+import soma.haeya.edupi_user.domain.Member;
+import soma.haeya.edupi_user.dto.request.MemberLoginRequest;
 
 @Component
 @HttpExchange("/api/v1/member")
