+++ src/main/java/soma/haeya/edupi_user/controller/MemberController.java
@@ -1,9 +1,13 @@
-package soma.haeya.edupi_user.login.controller;
+package soma.haeya.edupi_user.controller;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,9 +15,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import soma.haeya.edupi_user.login.dto.MemberLoginRequest;
-import soma.haeya.edupi_user.login.dto.TokenInfo;
-import soma.haeya.edupi_user.login.service.MemberService;
+import org.springframework.web.client.RestClient;
+import soma.haeya.edupi_user.dto.request.MemberLoginRequest;
+import soma.haeya.edupi_user.dto.TokenInfo;
+import soma.haeya.edupi_user.dto.request.SignupRequest;
+import soma.haeya.edupi_user.dto.response.ErrorResponse;
+import soma.haeya.edupi_user.service.MemberService;
+import soma.haeya.edupi_user.exception.DbValidException;
 
 @RestController
 @RequiredArgsConstructor
@@ -43,5 +51,20 @@ public class MemberController {
         return ResponseEntity.ok(tokenInfo);
     }
 
+    @PostMapping(value = "/signup")
+    public ResponseEntity<Void> createPost(@Valid @RequestBody SignupRequest signupRequest) {
+        // DB에 저장하기
+        RestClient restClient = RestClient.create();
 
+        return restClient.post()
+                .uri("http://localhost:8081/api/v1/user/save/signup")
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(signupRequest)
+                .retrieve()
+                .onStatus(HttpStatusCode::is4xxClientError, (req, res) -> {
+                    ErrorResponse errorResponse =  new ObjectMapper().readValue (res.getBody(), ErrorResponse.class);
+                    throw new DbValidException(errorResponse.message());
+                })
+                .toBodilessEntity();
+    }
 }
