+++ src/main/java/soma/haeya/edupi_user/dto/response/ErrorResponse.java
@@ -1,0 +1,4 @@
+package soma.haeya.edupi_user.dto.response;
+
+public record ErrorResponse(String message) {
+}
