--- src/main/java/soma/haeya/edupi_user/signup/exception/DbValidException.java
@@ -1,14 +1,0 @@
-package soma.haeya.edupi_user.signup.exception;
-
-public class DbValidException extends RuntimeException{
-    private int errorCode;
-
-    public DbValidException(String message) {
-        super(message);
-    }
-
-    public DbValidException(String message, int errorCode) {
-        super(message);
-        this.errorCode = errorCode;
-    }
-}
